Reworked several small things.
Main important change: switched to slist and *prepend* elements to the
global LRU queue (instead of to the tail as before).

Sanity checking against the python implementation, the numbers still
don't line up exactly.

The python simulator has been added for reference and its output has
been changed to be closer to the C++ version.