Few changes to defaults, fix slab_multi dumps.
Memshare dumps happened once per trace-hour. The slab dumps were only
advancing the last dump time to the timestamp of the request that
triggered the dump. This was causing small differences in when dumps
happened, which makes the last dumps accesses not match up.

Changed slab to match memshare for consistency.